qbfsat: Clean up and refactor data structures into `qbfsat.h`.
